Add some keys to e_type munging
The Aruba class munges a bunch of the Entity MIB tables but the e_type table
was missing some keys, causing ND1 to complain. This patch changes e_type to
merge the e_model table with the aruba_ap_type table, so the APs in e_type have
their correct type and the non-AP parts have their model (which looks like a
type and is better than nothing).