Stormshield: flip i_ ifindex logic projection
Mapping many tables into the STORMSHIELD-IF-MIB index space turned
out to be very verbose, and that was just for the i_ methods, let
alone all the lldp_, mac-address-table etc. linkage that might still
be coming.

Instead, returned to interfaces() returning ifEntry indexes, and
then tacking on some details like snsifUserName and snsifType.